+++ src/test/java/testsDominio/TestEnemigosExperiencia.java
@@ -1,4 +1,4 @@
-package tests_dominio;
+package testsDominio;
 
 import org.junit.Assert;
 import org.junit.Test;
