+++ src/test/java/testsDominio/TestPersonajeNuevo.java
@@ -1,4 +1,4 @@
-package tests_dominio;
+package testsDominio;
 
 import java.util.HashMap;
